Show a hint to sign up to guest users when they try to navigate to their personal space
